rework filenames, h/t Mike!
